Moved old log sample out to its own doc file.
