+++ Examples/StereoKitTest/Program.cs
@@ -24,11 +24,12 @@ class Program
 	static Matrix floorTr;
 	static Pose   windowDemoPose = new Pose();
 	static bool   windowDemoShow = false;
-	static bool   windowConsoleShow = false;
 	static Sprite powerButton;
 
 	static List<string> demoNames    = new List<string>();
 	static float        demoWinWidth = 50 * U.cm;
+
+	public static LogWindow logWindow;
 
 	public static bool WindowDemoShow
 	{
@@ -46,12 +47,12 @@ class Program
 
 	public static bool WindowConsoleShow
 	{
-		get => windowConsoleShow;
+		get => logWindow.Enabled;
 		set {
-			if (windowConsoleShow == value) return;
-			windowConsoleShow = value;
-			if (windowConsoleShow)
-				logPose = UI.PopupPose();
+			if (logWindow.Enabled == value) return;
+			logWindow.Enabled = value;
+			if (logWindow.Enabled)
+				logWindow.pose = UI.PopupPose();
 		}
 	}
 
@@ -78,16 +79,10 @@ class Program
 		SK.PreLoadLibrary();
 		Time.Scale = Tests.IsTesting ? 0 : 1;
 
-		/// :CodeSample: Log.Subscribe Log
-		/// Then you add the OnLog method into the log events like this in
-		/// your initialization code!
-		Log.Subscribe(OnLog);
-		/// :End:
-		
 		SK.AddStepper<PassthroughFBExt>();
-
-		SK.AddStepper<LogWindow>();
 		//SK.AddStepper<Win32PerformanceCounterExt>();
+		logWindow = SK.AddStepper<LogWindow>();
+		logWindow.Enabled = false;
 
 		// Initialize StereoKit
 		if (!SK.Initialize(settings))
@@ -164,15 +159,6 @@ class Program
 			return;
 
 		WindowDemoStep();
-
-		if (windowConsoleShow)
-		{
-			/// :CodeSample: Log.Subscribe Log
-			/// And in your Update loop, you can draw the window.
-			LogWindow();
-			/// And that's it!
-			/// :End:
-		}
 	}
 
 	static void WindowDemoStep()
@@ -212,49 +198,6 @@ class Program
 			SK.Quit();
 		UI.WindowEnd();
 	}
-
-	//////////////////////
-	// Debug log window //
-	//////////////////////
-
-	/// :CodeSample: Log.Subscribe Log
-	/// ### An in-application log window
-	/// Here's an example of using the Log.Subscribe method to build a simple
-	/// logging window. This can be pretty handy to have around somewhere in
-	/// your application!
-	/// 
-	/// Here's the code for the window, and log tracking.
-	static Pose         logPose   = new Pose(0, -0.1f, 0.5f, Quat.LookDir(Vec3.Forward));
-	static List<string> logList   = new List<string>();
-	static float        logIndex  = 0;
-	static string       logString = "";
-	static void OnLog(LogLevel level, string text)
-	{
-		logList.Insert(0, text.Length < 100 ? text + "\n" : text.Substring(0, 100) + "...\n");
-		UpdateLogStr((int)logIndex);
-	}
-
-	static void UpdateLogStr(int index)
-	{
-		logIndex  = Math.Max(Math.Min(index, logList.Count-1), 0);
-		logString = "";
-		for (int i = index; i < index + 15 && i < logList.Count; i++)
-			logString += logList[i];
-	}
-
-	static void LogWindow()
-	{
-		UI.WindowBegin("Log", ref logPose, new Vec2(40, 0) * U.cm);
-
-		UI.LayoutPushCut(UICut.Right, UI.LineHeight);
-		if (UI.VSlider("scroll", ref logIndex, 0, Math.Max(logList.Count - 3, 0), 1))
-			UpdateLogStr((int)logIndex);
-		UI.LayoutPop();
-
-		UI.Text(logString);
-		UI.WindowEnd();
-	}
-	/// :End:
 
 	/// :CodeSample: AppFocus SK.AppFocus
 	/// ### Checking for changes in application focus
